+++ cmd/setup/pubsub/base.go
@@ -13,7 +13,6 @@ import (
 	"github.com/IsaacDSC/gqueue/internal/fetcher"
 	"github.com/IsaacDSC/gqueue/internal/interstore"
 	"github.com/IsaacDSC/gqueue/internal/storests"
-	"github.com/IsaacDSC/gqueue/pkg/ctxlogger"
 	"github.com/IsaacDSC/gqueue/pkg/pubadapter"
 	"github.com/googleapis/gax-go/v2"
 	"google.golang.org/grpc/codes"
@@ -89,24 +88,7 @@ func (s *Service) Start(ctx context.Context, env cfg.Config) {
 	s.memStore.LoadInMemStore(ctx)
 
 	// task refresh mem store
-	go func() {
-		l := ctxlogger.GetLogger(ctx)
-		trigger := time.NewTicker(time.Minute)
-		for {
-			select {
-			case <-trigger.C:
-				if err := s.memStore.LoadInMemStore(ctx); err != nil {
-					l.Error("Error refreshing mem store with events from persistent store", "error", err)
-					continue
-				}
-
-				l.Info("Executed periodic refresh of mem store with events from persistent store", "scope", "pubsub")
-			case <-ctx.Done():
-				trigger.Stop()
-				return
-			}
-		}
-	}()
+	go s.syncMemStore(ctx)
 
 	s.server = s.startHttpServer(ctx, env)
 }
